+++ terrawrap/exceptions.py
@@ -9,3 +9,7 @@ class NotTerraformConfigDirectory(RuntimeError):
 
 class NoDependency(Exception):
     """Error raised when processing a directory that contains .tf_wrapper config files with no dependency"""
+
+
+class EnvVarResolutionError(RuntimeError):
+    """Error raised when a .tf_wrapper envvar cannot be resolved to a value"""